magit-show-refs: Match documented *-use-buffer-arguments behavior
The manual says that magit-show-refs (1) behaves as a refresh prefix
when called from a magit-refs-mode buffer, (2) respects
magit-prefix-use-buffer-arguments when called as a transient, and (3)
respects magit-direct-use-buffer-arguments otherwise.  The third part
is not the case.

Match the described behavior by updating magit-show-refs-arguments to
look at magit-direct-use-buffer-arguments unless the magit-show-refs
transient overrides the value with magit-prefix-use-buffer-arguments.