magit-wip-commit-index: Call read-tree with -i
When calling git-read-tree to create a temporary index for the working
tree wip ref, we perform a merge in addition to a read [0].  The
advantage of doing a merge---which has been done since 59e43cbe7
(address issues concerning `git read-tree', 2015-09-11)---is
performance.  When a merge is specified for a single tree, as is the
case here, it means "copy over the index's stat info for paths that
match between the tree and the index".  So, our newly constructed
temporary index is populated with the stat info from the main index
for unchanged paths, and we benefit from Git's stat comparison
optimization.

However, doing a merge means that we a susceptible to a read-tree
failure that reports

    error: Entry '<file>' not uptodate. Cannot merge

when some of the index entries are marked with "assume unchanged" or
"skip worktree" bits.  In particular, the problem arises when one of
the marked entries in the main index differs from a file in the wip
working tree ref.  The easiest way for this to happen is (1) the user
changes and saves an unmarked file in the working tree, (2) marks the
file with "assume unchanged" or "skip worktree", and (3) changes and
saves the file again.

Pass -i to read-tree, which disables the up-to-date check.  This
should be fine to do given that we're writing to a temporary index and
that, as explained above, we're using a single-tree "merge" rather
than the merge of HEAD and another tree for which the check is
relevant.

Re: 4037

[0]: We do so with --reset rather than -m, but we specify only a
     single tree, so the behavior is the same.